inigt

+++ frontend/modules/Wechat/controllers/DefaultController.php
@@ -227,7 +227,7 @@ class DefaultController extends Controller
         $role = yii::$app->request->get('role', MgUsers::IS_PLAYER );
         $subs = MgUsers::findAll( [ 'id'=>$ids,'is_bd'=>$role ] );
         //获取活跃用户数
-        $score = yii::$app->redis->ZSCORE( 'ACT_ZSET_RANK', "user_act_users_{$mgInfo->id}" );       
+        $score = (int)yii::$app->redis->ZSCORE( 'ACT_ZSET_RANK', "user_act_users_{$mgInfo->id}" );       
         return $this->render('my_friend', [
             'subs'=> $subs,
             'role'=> $role,
